eliquid add page

+++ src/main/java/com/tiger8bit/vapeshop/repository/BlendRatioRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BlendRatioRepository extends CrudRepository<BlendRatio, Long> {
+public interface BlendRatioRepository extends CrudRepository<BlendRatio, Integer> {
 
 }
