+++ src/main/java/com/tiger8bit/vapeshop/repository/BrandRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BrandRepository extends CrudRepository<Brand, Long> {
+public interface BrandRepository extends CrudRepository<Brand, Integer> {
 
 }
