+++ src/main/java/com/tiger8bit/vapeshop/repository/DeliveryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DeliveryRepository extends CrudRepository<Delivery, Long> {
+public interface DeliveryRepository extends CrudRepository<Delivery, Integer> {
 
 }
