+++ src/main/java/com/tiger8bit/vapeshop/repository/ELiquidRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ELiquidRepository extends CrudRepository<ELiquid, Long> {
+public interface ELiquidRepository extends CrudRepository<ELiquid, Integer> {
 
 }
