+++ src/main/java/com/tiger8bit/vapeshop/repository/PhoneNumberRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PhoneNumberRepository extends CrudRepository<PhoneNumber, Long> {
+public interface PhoneNumberRepository extends CrudRepository<PhoneNumber, Integer> {
 
 }
