+++ src/main/java/com/tiger8bit/vapeshop/repository/PriceRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PriceRepository extends CrudRepository<Price, Long> {
+public interface PriceRepository extends CrudRepository<Price, Integer> {
 
 }
