+++ src/main/java/com/tiger8bit/vapeshop/repository/ProductImageRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ProductImageRepository extends CrudRepository<ProductImage, Long> {
+public interface ProductImageRepository extends CrudRepository<ProductImage, Integer> {
 
 }
