+++ src/main/java/com/tiger8bit/vapeshop/repository/SaleRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SaleRepository extends CrudRepository<Sale, Long> {
+public interface SaleRepository extends CrudRepository<Sale, Integer> {
 
 }
