+++ src/main/java/com/tiger8bit/vapeshop/repository/VapeshopRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface VapeshopRepository extends CrudRepository<Vapeshop, Long> {
+public interface VapeshopRepository extends CrudRepository<Vapeshop, Integer> {
 
 }
