+++ src/main/java/com/tiger8bit/vapeshop/service/ELiquidServiceImpl.java
@@ -22,12 +22,12 @@ public class ELiquidServiceImpl implements ELiquidService {
     }
 
     @Override
-    public void deleteByID(long id) {
+    public void deleteByID(int id) {
         repository.deleteById(id);
     }
 
     @Override
-    public ELiquid findByID(long id) {
+    public ELiquid findByID(int id) {
         return repository.findById(id).get();
     }
 }
