+++ src/main/java/com/tiger8bit/vapeshop/service/FlavorProfileServiceImpl.java
@@ -22,12 +22,12 @@ public class FlavorProfileServiceImpl implements FlavorProfileService {
     }
 
     @Override
-    public void deleteByID(long id) {
+    public void deleteByID(int id) {
         repository.deleteById(id);
     }
 
     @Override
-    public FlavorProfile findByID(long id) {
+    public FlavorProfile findByID(int id) {
         return repository.findById(id).get();
     }
 }
