+++ src/main/java/com/tiger8bit/vapeshop/service/PhoneNumberService.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface PhoneNumberService {
     List<PhoneNumber> findAll();
     PhoneNumber save(PhoneNumber value);
-    void deleteByID(long id);
-    PhoneNumber findByID(long id);
+    void deleteByID(int id);
+    PhoneNumber findByID(int id);
 }
