+++ src/main/java/com/tiger8bit/vapeshop/service/ProductService.java
@@ -5,6 +5,6 @@ import java.util.List;
 public interface ProductService {
     List<Product> findAll();
     Product save(Product value);
-    void deleteByID(long id);
-    Product findByID(long id);
+    void deleteByID(int id);
+    Product findByID(int id);
 }
